Option B: Keep double public API, QuantityValue internal
Revert public-facing API from QuantityValue to double while keeping
QuantityValue internally for the conversion pipeline.

Changes:
- IQuantity.Value, From(), Create() → double
- IArithmeticQuantity/ILogarithmicQuantity operator interfaces → double
- Generated quantities: _value field, properties, operators, factories → double
- QuantityExtensions.As() → returns double
- Quantity static methods → double
- UnitConverter static convenience methods → double
- Comparison.cs → double
- CodeGen templates updated, code regenerated
- Test/benchmark code updated for double types

Build: 0 library errors, 0 warnings.
Tests: 2509 failures out of 51899 — primarily ToUnit roundtrip tests
that relied on QuantityValue exact rational precision for lossless
conversion, which is lost when storing values as double.

This demonstrates the fundamental trade-off: Option B loses conversion
roundtrip precision that was a key benefit of the QuantityValue PR.

+++ CodeGen/Generators/UnitsNetGen/UnitTestBaseClassGenerator.cs
@@ -599,7 +599,7 @@ namespace UnitsNet.Tests
             }
 
             Writer.WL($@"
-        public void Parse(string culture, string quantityString, {_unitEnumName} expectedUnit, decimal expectedValue)
+        public void Parse(string culture, string quantityString, {_unitEnumName} expectedUnit, double expectedValue)
         {{
             using var _ = new CultureScope(culture);
             var parsed = {_quantity.Name}.Parse(quantityString);
@@ -644,7 +644,7 @@ namespace UnitsNet.Tests
             }
 
             Writer.WL($@"
-        public void TryParse(string culture, string quantityString, {_unitEnumName} expectedUnit, decimal expectedValue)
+        public void TryParse(string culture, string quantityString, {_unitEnumName} expectedUnit, double expectedValue)
         {{
             using var _ = new CultureScope(culture);
             Assert.True({_quantity.Name}.TryParse(quantityString, out {_quantity.Name} parsed));
@@ -1129,8 +1129,8 @@ namespace UnitsNet.Tests
             var quantity = {_quantity.Name}.From{_baseUnit.PluralName}(firstValue);
             var otherQuantity = {_quantity.Name}.From{_baseUnit.PluralName}(secondValue);
             {differenceResultType} maxTolerance = quantity > otherQuantity ? quantity - otherQuantity : otherQuantity - quantity;
-            var largerTolerance = maxTolerance * 1.1m;
-            var smallerTolerance = maxTolerance / 1.1m;
+            var largerTolerance = maxTolerance * 1.1;
+            var smallerTolerance = maxTolerance / 1.1;
             Assert.True(quantity.Equals(quantity, {differenceResultType}.Zero));
             Assert.True(quantity.Equals(quantity, maxTolerance));
             Assert.True(quantity.Equals(otherQuantity, largerTolerance));
@@ -1162,8 +1162,8 @@ namespace UnitsNet.Tests
             var quantity = {_quantity.Name}.From{_baseUnit.PluralName}(firstValue);
             var otherQuantity = {_quantity.Name}.From{_baseUnit.PluralName}(secondValue);
             {differenceResultType} maxTolerance = quantity > otherQuantity ? quantity - otherQuantity : otherQuantity - quantity;
-            var largerTolerance = maxTolerance * 1.1m;
-            var smallerTolerance = maxTolerance / 1.1m;
+            var largerTolerance = maxTolerance * 1.1;
+            var smallerTolerance = maxTolerance / 1.1;
             Assert.True(quantity.Equals(quantity, {differenceResultType}.Zero));
             Assert.True(quantity.Equals(quantity, maxTolerance));
             Assert.True(quantity.Equals(otherQuantity, maxTolerance));
